// frontend/src/features/commons/types.ts
export type ReferenceType = "case" | "patient" | "channel";

export interface ObjectReference {
  id: number;
  message_id: number;
  referenceable_type: ReferenceType;
  referenceable_id: number;
  display_name: string;
}